+++ plugin/sampling/strategystore/adaptive/factory_test.go
@@ -15,16 +15,23 @@
 package adaptive
 
 import (
+	"errors"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
 	"github.com/uber/jaeger-lib/metrics"
 	"go.uber.org/zap"
 
+	"github.com/jaegertracing/jaeger/cmd/collector/app/sampling/model"
 	ss "github.com/jaegertracing/jaeger/cmd/collector/app/sampling/strategystore"
 	"github.com/jaegertracing/jaeger/pkg/config"
+	"github.com/jaegertracing/jaeger/pkg/distributedlock"
+	lmocks "github.com/jaegertracing/jaeger/pkg/distributedlock/mocks"
 	"github.com/jaegertracing/jaeger/plugin"
+	"github.com/jaegertracing/jaeger/storage/samplingstore"
+	smocks "github.com/jaegertracing/jaeger/storage/samplingstore/mocks"
 )
 
 var _ ss.Factory = new(Factory)
@@ -61,7 +68,70 @@ func TestFactory(t *testing.T) {
 	assert.Equal(t, time.Second, f.options.LeaderLeaseRefreshInterval)
 	assert.Equal(t, time.Second*2, f.options.FollowerLeaseRefreshInterval)
 
-	assert.NoError(t, f.Initialize(metrics.NullFactory, zap.NewNop()))
-	_, err := f.CreateStrategyStore()
+	assert.NoError(t, f.Initialize(metrics.NullFactory, &mockSamplingStoreFactory{}, zap.NewNop()))
+	_, _, err := f.CreateStrategyStore()
 	assert.NoError(t, err)
+}
+
+func TestBadConfigFail(t *testing.T) {
+	tests := []string{
+		"--sampling.aggregation-buckets=0",
+		"--sampling.calculation-interval=0",
+		"--sampling.buckets-for-calculation=0",
+	}
+
+	for _, tc := range tests {
+		f := NewFactory()
+		v, command := config.Viperize(f.AddFlags)
+		command.ParseFlags([]string{
+			tc,
+		})
+
+		f.InitFromViper(v, zap.NewNop())
+
+		assert.NoError(t, f.Initialize(metrics.NullFactory, &mockSamplingStoreFactory{}, zap.NewNop()))
+		_, _, err := f.CreateStrategyStore()
+		assert.Error(t, err)
+	}
+}
+
+func TestSamplingStoreFactoryFails(t *testing.T) {
+	f := NewFactory()
+
+	// nil fails
+	assert.Error(t, f.Initialize(metrics.NullFactory, nil, zap.NewNop()))
+
+	// fail if lock fails
+	assert.Error(t, f.Initialize(metrics.NullFactory, &mockSamplingStoreFactory{lockFailsWith: errors.New("fail")}, zap.NewNop()))
+
+	// fail if store fails
+	assert.Error(t, f.Initialize(metrics.NullFactory, &mockSamplingStoreFactory{storeFailsWith: errors.New("fail")}, zap.NewNop()))
+}
+
+type mockSamplingStoreFactory struct {
+	lockFailsWith  error
+	storeFailsWith error
+}
+
+func (m *mockSamplingStoreFactory) CreateLock() (distributedlock.Lock, error) {
+	if m.lockFailsWith != nil {
+		return nil, m.lockFailsWith
+	}
+
+	mockLock := &lmocks.Lock{}
+	mockLock.On("Acquire", mock.Anything, mock.Anything).Return(true, nil)
+
+	return mockLock, nil
+}
+func (m *mockSamplingStoreFactory) CreateSamplingStore() (samplingstore.Store, error) {
+	if m.storeFailsWith != nil {
+		return nil, m.storeFailsWith
+	}
+
+	mockStorage := &smocks.Store{}
+	mockStorage.On("GetLatestProbabilities").Return(make(model.ServiceOperationProbabilities), nil)
+	mockStorage.On("GetThroughput", mock.AnythingOfType("time.Time"), mock.AnythingOfType("time.Time")).
+		Return([]*model.Throughput{}, nil)
+
+	return mockStorage, nil
 }