+++ plugin/sampling/strategystore/factory_test.go
@@ -18,6 +18,7 @@ package strategystore
 import (
 	"errors"
 	"flag"
+	"os"
 	"testing"
 
 	"github.com/spf13/viper"
@@ -27,52 +28,79 @@ import (
 	"go.uber.org/zap"
 
 	ss "github.com/jaegertracing/jaeger/cmd/collector/app/sampling/strategystore"
+	"github.com/jaegertracing/jaeger/pkg/distributedlock"
 	"github.com/jaegertracing/jaeger/plugin"
+	"github.com/jaegertracing/jaeger/storage"
+	"github.com/jaegertracing/jaeger/storage/samplingstore"
 )
+
+func clearEnv() {
+	os.Setenv(SamplingTypeEnvVar, "static")
+}
 
 var _ ss.Factory = new(Factory)
 var _ plugin.Configurable = new(Factory)
 
 func TestNewFactory(t *testing.T) {
-	f, err := NewFactory(FactoryConfig{StrategyStoreType: staticStrategyStoreType})
-	require.NoError(t, err)
-	assert.NotEmpty(t, f.factories)
-	assert.NotEmpty(t, f.factories[staticStrategyStoreType])
-	assert.Equal(t, staticStrategyStoreType, f.StrategyStoreType)
+	tests := []struct {
+		strategyStoreType string
+		expectError       bool
+	}{
+		{
+			strategyStoreType: "static",
+		},
+		{
+			strategyStoreType: "adaptive",
+		},
+		{
+			strategyStoreType: "nonsense",
+			expectError:       true,
+		},
+	}
 
-	mock := new(mockFactory)
-	f.factories[staticStrategyStoreType] = mock
+	mockSSFactory := &mockSamplingStoreFactory{}
 
-	assert.NoError(t, f.Initialize(metrics.NullFactory, zap.NewNop()))
-	_, err = f.CreateStrategyStore()
-	assert.NoError(t, err)
+	for _, tc := range tests {
+		f, err := NewFactory(FactoryConfig{StrategyStoreType: Kind(tc.strategyStoreType)})
+		if tc.expectError {
+			assert.Error(t, err)
+			continue
+		}
+		assert.NotEmpty(t, f.factories)
+		assert.NotEmpty(t, f.factories[Kind(tc.strategyStoreType)])
+		assert.Equal(t, Kind(tc.strategyStoreType), f.StrategyStoreType)
 
-	// force the mock to return errors
-	mock.retError = true
-	assert.EqualError(t, f.Initialize(metrics.NullFactory, zap.NewNop()), "error initializing store")
-	_, err = f.CreateStrategyStore()
-	assert.EqualError(t, err, "error creating store")
+		mock := new(mockFactory)
+		f.factories[Kind(tc.strategyStoreType)] = mock
 
-	f.StrategyStoreType = "nonsense"
-	_, err = f.CreateStrategyStore()
-	assert.EqualError(t, err, "no nonsense strategy store registered")
+		assert.NoError(t, f.Initialize(metrics.NullFactory, mockSSFactory, zap.NewNop()))
+		_, _, err = f.CreateStrategyStore()
+		require.NoError(t, err)
 
-	_, err = NewFactory(FactoryConfig{StrategyStoreType: "nonsense"})
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "unknown sampling strategy store type")
+		// force the mock to return errors
+		mock.retError = true
+		assert.EqualError(t, f.Initialize(metrics.NullFactory, mockSSFactory, zap.NewNop()), "error initializing store")
+		_, _, err = f.CreateStrategyStore()
+		assert.EqualError(t, err, "error creating store")
+
+		// request something that doesn't exist
+		f.StrategyStoreType = "doesntexist"
+		_, _, err = f.CreateStrategyStore()
+		assert.EqualError(t, err, "no doesntexist strategy store registered")
+	}
 }
 
 func TestConfigurable(t *testing.T) {
 	clearEnv()
 	defer clearEnv()
 
-	f, err := NewFactory(FactoryConfig{StrategyStoreType: staticStrategyStoreType})
+	f, err := NewFactory(FactoryConfig{StrategyStoreType: "static"})
 	require.NoError(t, err)
 	assert.NotEmpty(t, f.factories)
-	assert.NotEmpty(t, f.factories[staticStrategyStoreType])
+	assert.NotEmpty(t, f.factories["static"])
 
 	mock := new(mockFactory)
-	f.factories[staticStrategyStoreType] = mock
+	f.factories["static"] = mock
 
 	fs := new(flag.FlagSet)
 	v := viper.New()
@@ -100,16 +128,25 @@ func (f *mockFactory) InitFromViper(v *viper.Viper, logger *zap.Logger) {
 	f.logger = logger
 }
 
-func (f *mockFactory) CreateStrategyStore() (ss.StrategyStore, error) {
+func (f *mockFactory) CreateStrategyStore() (ss.StrategyStore, ss.Aggregator, error) {
 	if f.retError {
-		return nil, errors.New("error creating store")
+		return nil, nil, errors.New("error creating store")
 	}
-	return nil, nil
+	return nil, nil, nil
 }
 
-func (f *mockFactory) Initialize(metricsFactory metrics.Factory, logger *zap.Logger) error {
+func (f *mockFactory) Initialize(metricsFactory metrics.Factory, ssFactory storage.SamplingStoreFactory, logger *zap.Logger) error {
 	if f.retError {
 		return errors.New("error initializing store")
 	}
 	return nil
+}
+
+type mockSamplingStoreFactory struct{}
+
+func (m *mockSamplingStoreFactory) CreateLock() (distributedlock.Lock, error) {
+	return nil, nil
+}
+func (m *mockSamplingStoreFactory) CreateSamplingStore() (samplingstore.Store, error) {
+	return nil, nil
 }