+++ plugin/storage/cassandra/factory_test.go
@@ -100,6 +100,12 @@ func TestCassandraFactory(t *testing.T) {
 	_, err = f.CreateArchiveSpanWriter()
 	assert.NoError(t, err)
 
+	_, err = f.CreateLock()
+	assert.NoError(t, err)
+
+	_, err = f.CreateSamplingStore()
+	assert.NoError(t, err)
+
 	assert.NoError(t, f.Close())
 }